+++ apps/windows/RecApp/Shell/CabinetWindow.h
@@ -2,6 +2,10 @@
 
 #include "../Web/WebView2Host.h"
 
+#ifdef _WIN32
+#include <winrt/Microsoft.UI.Xaml.Controls.h>
+#endif
+
 #include <string_view>
 
 namespace graf::windows {
@@ -13,6 +17,10 @@ public:
     [[nodiscard]] RouteEvaluation openCabinet();
     [[nodiscard]] RouteEvaluation open(std::string_view url);
     [[nodiscard]] WebView2Host& webView() noexcept { return webView_; }
+
+#ifdef _WIN32
+    void attach(Microsoft::UI::Xaml::Controls::WebView2 control) { webView_.attach(control); }
+#endif
 
 private:
     WebView2Host webView_;